peer, connector는 매개변수로 넘길때 ({peer, connector}) 늘 같이 따라다니네요. 이렇게 자주 매개변수로 함께 넘겨서 메소드콜한다면 peer와 connector를 하나로 묶고 해당메소드를 webrtc에서 묶은 클래스로 위임해야하지 않을까합니다.
일단 안쓰는 멤버변수가 많은 connector를 제거하고 peer와 통합했습니다.
코드가 동작할려면 다른쪽도 같이 고쳐야하지만 클래스통합은 비교적 큰작업이니 일단 나중에...
