+++ src/Peer.js
@@ -1,14 +1,60 @@
 import Emitter from 'event-emitter';
 
 export default class Peer {
-  constructor(id) {
+  constructor(id, peer) {
     this.id = id;
+    this._peer = peer;
     this._dc = null;
     this._emitter = new Emitter();
     this.localSdp = null;
     this.remoteSdp = null;
     this.localStream = null;
     this.remoteStream = null;
+  }
+ 
+  set onIceCandidate(func) {
+    this._peer.onicecandidate = func;
+  }
+
+  set onTrack(func) {
+    this._peer.ontrack = func;
+  }
+
+  set onDataChannel(func) {
+    this._peer.ondatachannel = func;
+  }
+
+  createOffer() {
+    return this._peer.createOffer();
+  }
+
+  createAnswer() {
+    return this._peer.createAnswer();
+  }
+
+  createDataChannel(channelName) {
+    if (!this._peer.createDataChannel) return;
+    return this._peer.createDataChannel(channelName);
+  }
+
+  addTrack(stream) {
+    if (!stream) return;
+
+    stream.getTracks().forEach(track => {
+      this._peer.addTrack(track, stream);
+    });
+  }
+
+  setLocalDescription(sdp) {
+    return this._peer.setLocalDescription(sdp);
+  }
+
+  setRemoteDescription(sdp) {
+    return this._peer.setRemoteDescription(new RTCSessionDescription(sdp));
+  }
+
+  addIceCandidate(candidate) {
+    return this._peer.addIceCandidate(candidate);
   }
 
   on(eventName, listener) {
@@ -32,7 +78,7 @@ export default class Peer {
 
   setDataChannel(dc) {
     this._dc = dc;
-    
+
     dc.onmessage = ({ data }) => this._emitter.emit('message', data);
     dc.onclose = () => this._emitter.emit('close');
     dc.onopen = () => this._emitter.emit('open');
